project-2_todo-list-full-stack-nextjs-fastapi/project_1/server/pyproject.toml
[tool.poetry]
name = "todo-list"
version = "0.1.0"
description = ""
authors = ["Usama Qazi <usamyharis@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
sqlmodel = "^0.0.16"
python-dotenv = "^1.0.1"
uvicorn = "^0.29.0"
psycopg2 = "^2.9.9"
pytest = "^8.1.1"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
